obtener coincidencias

+++ database/sc.sql
@@ -11,7 +11,7 @@ CREATE TABLE users(
 CREATE TABLE indicators(
     id_indi INT(11) NOT NULL PRIMARY KEY AUTO_INCREMENT,
     symbol VARCHAR(10) NOT NULL,
-    current_price DECIMAL(15,15) NOT NULL,
+    current_price FLOAT(15,15) NOT NULL,
     indi_username INT(11) NOT NULL,
     created_at timestamp NOT NULL DEFAULT current_timestamp,
     CONSTRAINT fk_indic FOREIGN KEY(indi_username) REFERENCES users(id)
